requirements.txt
flask
qrcode
pillow
pywebview
